Add 32 bit builds

+++ script/build_docker.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set +e
 . script/env_deploy.sh
 set -e
 
@@ -8,7 +9,7 @@ then
  export DEST="$1"
 fi
 
-if [[ -d download-artifact ]]
+if [[ -d download-artifact && "$SKIP_BUILD_GO" != yes ]]
 then
 (
  cd download-artifact
@@ -18,26 +19,41 @@ then
 ) ||:
 fi
 
-if [[ -f $DEPLOYMENT/$archive_standalone.tar.xz ]]
+echo $archive_standalone
+
+if [[ -f $DEPLOYMENT/$archive_standalone.tar.xz && "$SKIP_BUILD_GO" != yes ]]
 then
    pushd $DEPLOYMENT
    tar -xvf $archive_standalone.tar.xz
-   SRC_ROOT=$PWD/$archive_standalone
+   ls $PWD
+   echo $archive_standalone
+   export SRC_ROOT=$PWD/$archive_standalone
 #   BUILD="$SRC_ROOT/build"
-   GOFLAGS="-mod=vendor $GOFLAGS"
-   VERSION_SINGBOX="$(cat $SRC_ROOT/SingBox.Version)"
-   LAST_ACTION='rm -rf "$SRC_ROOT"'
+   export GOFLAGS="-mod=vendor $GOFLAGS"
+   export VERSION_SINGBOX="$(cat $SRC_ROOT/SingBox.Version)"
+   export LAST_ACTION='rm -rf "$SRC_ROOT"'
    popd
+   
 else
   LAST_ACTION="echo fine"
 fi 
 
+(
+echo "$SRC_ROOT"
+cd "$SRC_ROOT"
 
-( . script/build_go.sh; )
+if [[ "$SKIP_BUILD_GO" != yes ]]
+then
+(
+. script/build_go.sh; 
+)
+fi
 
 cmake -S $SRC_ROOT -B "$BUILD" -GNinja -DNKR_DEFAULT_VERSION="${INPUT_VERSION:-5.0.0}"
 cmake --build "$BUILD" -v -j $(nproc)
-
-( . script/deploy_linux64.sh; )
+(
+. script/deploy_linux64.sh; 
+)
+)
 
 eval "$LAST_ACTION"
